crypto_analysis_project/requirements.txt
openai>=1.0.0
langchain>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pybit>=5.5.0
pandas>=2.0.0
aiohttp>=3.9.0